# openapi.json
{
  "components": {
    "schemas": {
      "AnalysisStatus": {
        "enum": [
          "not_requested",
          "pending",
          "completed",
          "failed"
        ],
        "title": "AnalysisStatus",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LeadCreate": {
        "example": {
          "budget_text": "around $20k",
          "company": "Acme Web Co",
          "deadline_text_input": "within 2 months",
          "description": "We need a marketing website redesign with a CMS and blog.",
          "email": "alice@example.com",
          "name": "Alice Anderson",
          "source": "website"
        },
        "properties": {
          "budget_text": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Text"
          },
          "company": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "deadline_text_input": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline Text Input"
          },
          "description": {
            "maxLength": 5000,
            "minLength": 20,
            "title": "Description",
            "type": "string"
          },
          "email": {
            "format": "email",
            "maxLength": 254,
            "title": "Email",
            "type": "string"
          },
          "name": {
            "maxLength": 120,
            "minLength": 2,
            "title": "Name",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/LeadSource",
            "default": "manual"
          }
        },
        "required": [
          "name",
          "email",
          "description"
        ],
        "title": "LeadCreate",
        "type": "object"
      },
      "LeadListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LeadRead"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "LeadListResponse",
        "type": "object"
      },
      "LeadPriority": {
        "enum": [
          "low",
          "medium",
          "high",
          "urgent"
        ],
        "title": "LeadPriority",
        "type": "string"
      },
      "LeadRead": {
        "properties": {
          "ai_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Model"
          },
          "ai_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Summary"
          },
          "analysis_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Error"
          },
          "analysis_reasons": {
            "items": {
              "type": "string"
            },
            "title": "Analysis Reasons",
            "type": "array"
          },
          "analysis_status": {
            "$ref": "#/components/schemas/AnalysisStatus"
          },
          "budget_max": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Max"
          },
          "budget_min": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Min"
          },
          "budget_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Text"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Currency"
          },
          "deadline_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline Text"
          },
          "deadline_text_input": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline Text Input"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "last_notified_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Notified At"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "notification_error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notification Error"
          },
          "notification_status": {
            "$ref": "#/components/schemas/NotificationStatus"
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LeadPriority"
              },
              {
                "type": "null"
              }
            ]
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "recommended_action": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommended Action"
          },
          "source": {
            "$ref": "#/components/schemas/LeadSource"
          },
          "status": {
            "$ref": "#/components/schemas/LeadStatus"
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "title": "Tags",
            "type": "array"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "company",
          "source",
          "budget_text",
          "deadline_text_input",
          "description",
          "status",
          "priority",
          "category",
          "ai_summary",
          "budget_min",
          "budget_max",
          "currency",
          "deadline_text",
          "recommended_action",
          "tags",
          "confidence",
          "analysis_reasons",
          "analysis_status",
          "analysis_error",
          "ai_model",
          "prompt_version",
          "notification_status",
          "notification_error",
          "last_notified_at",
          "created_at",
          "updated_at"
        ],
        "title": "LeadRead",
        "type": "object"
      },
      "LeadSortField": {
        "enum": [
          "created_at",
          "updated_at",
          "name",
          "priority",
          "status"
        ],
        "title": "LeadSortField",
        "type": "string"
      },
      "LeadSource": {
        "enum": [
          "website",
          "email",
          "telegram",
          "referral",
          "manual",
          "other"
        ],
        "title": "LeadSource",
        "type": "string"
      },
      "LeadStatus": {
        "enum": [
          "new",
          "contacted",
          "qualified",
          "won",
          "lost"
        ],
        "title": "LeadStatus",
        "type": "string"
      },
      "LeadUpdate": {
        "description": "Partial update; only provided fields are changed.",
        "example": {
          "priority": "high",
          "status": "contacted"
        },
        "properties": {
          "budget_text": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Budget Text"
          },
          "category": {
            "anyOf": [
              {
                "maxLength": 80,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "company": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "deadline_text_input": {
            "anyOf": [
              {
                "maxLength": 160,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deadline Text Input"
          },
          "description": {
            "anyOf": [
              {
                "maxLength": 5000,
                "minLength": 20,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "email": {
            "anyOf": [
              {
                "format": "email",
                "maxLength": 254,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "minLength": 2,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "priority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LeadPriority"
              },
              {
                "type": "null"
              }
            ]
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LeadSource"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LeadStatus"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "LeadUpdate",
        "type": "object"
      },
...
      "NotificationStatus": {
        "enum": [
          "not_required",
          "pending",
          "sent",
          "failed"
        ],
        "title": "NotificationStatus",
        "type": "string"
      },
...
      "SortOrder": {
        "enum": [
          "asc",
          "desc"
        ],
        "title": "SortOrder",
        "type": "string"
      },
      "TestTelegramRequest": {
        "properties": {
          "chat_id": {
            "anyOf": [
              {
                "maxLength": 64,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chat Id"
          }
        },
        "title": "TestTelegramRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "LeadFlow AI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/leads": {
      "get": {
        "operationId": "list_leads_api_v1_leads_get",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 160,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LeadStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "description": "Filter by priority; 'none' isolates unclassified leads.",
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LeadPriority"
                },
                {
                  "const": "none",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by priority; 'none' isolates unclassified leads.",
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LeadSource"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/LeadSortField",
              "default": "created_at"
            }
          },
          {
            "in": "query",
            "name": "sort_order",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "desc"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Leads",
        "tags": [
          "leads"
        ]
      },
      "post": {
        "operationId": "create_lead_api_v1_leads_post",
        "parameters": [
          {
            "description": "Run AI analysis after saving. AI failure still returns 201.",
            "in": "query",
            "name": "analyze",
            "required": false,
            "schema": {
              "default": false,
              "description": "Run AI analysis after saving. AI failure still returns 201.",
              "title": "Analyze",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Lead",
        "tags": [
          "leads"
        ]
      }
    },
    "/api/v1/leads/export": {
      "get": {
        "operationId": "export_leads_csv_api_v1_leads_export_get",
        "parameters": [
          {
            "in": "query",
            "name": "search",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 160,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LeadStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "description": "Filter by priority; 'none' isolates unclassified leads.",
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LeadPriority"
                },
                {
                  "const": "none",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by priority; 'none' isolates unclassified leads.",
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/LeadSource"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Source"
            }
          },
          {
            "in": "query",
            "name": "sort_by",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/LeadSortField",
              "default": "created_at"
            }
          },
          {
            "in": "query",
            "name": "sort_order",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/SortOrder",
              "default": "desc"
            }
          },
          {
            "description": "Prefix a UTF-8 BOM so Excel detects the encoding.",
            "in": "query",
            "name": "bom",
            "required": false,
            "schema": {
              "default": false,
              "description": "Prefix a UTF-8 BOM so Excel detects the encoding.",
              "title": "Bom",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Export Leads Csv",
        "tags": [
          "leads"
        ]
      }
    },
    "/api/v1/leads/{lead_id}": {
      "delete": {
        "operationId": "delete_lead_api_v1_leads__lead_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "lead_id",
            "required": true,
            "schema": {
              "title": "Lead Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Lead",
        "tags": [
          "leads"
        ]
      },
      "get": {
        "operationId": "get_lead_api_v1_leads__lead_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "lead_id",
            "required": true,
            "schema": {
              "title": "Lead Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Lead",
        "tags": [
          "leads"
        ]
      },
      "patch": {
        "operationId": "update_lead_api_v1_leads__lead_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "lead_id",
            "required": true,
            "schema": {
              "title": "Lead Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Lead",
        "tags": [
          "leads"
        ]
      }
    },
    "/api/v1/leads/{lead_id}/analyze": {
      "post": {
        "operationId": "analyze_existing_lead_api_v1_leads__lead_id__analyze_post",
        "parameters": [
          {
            "in": "path",
            "name": "lead_id",
            "required": true,
            "schema": {
              "title": "Lead Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Analyze Existing Lead",
        "tags": [
          "leads"
        ]
      }
    },
    "/api/v1/leads/{lead_id}/notify": {
      "post": {
        "operationId": "notify_existing_lead_api_v1_leads__lead_id__notify_post",
        "parameters": [
          {
            "in": "path",
            "name": "lead_id",
            "required": true,
            "schema": {
              "title": "Lead Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Notify Existing Lead",
        "tags": [
          "leads"
        ]
      }
    },
...
    "/api/v1/settings": {
      "get": {
        "operationId": "read_settings_api_v1_settings_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsRead"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Read Settings",
        "tags": [
          "settings"
        ]
      },
      "patch": {
        "operationId": "update_settings_api_v1_settings_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SettingsUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Settings",
        "tags": [
          "settings"
        ]
      }
    },
    "/api/v1/settings/test-telegram": {
      "post": {
        "operationId": "test_telegram_api_v1_settings_test_telegram_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TestTelegramRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestTelegramResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Test Telegram",
        "tags": [
          "settings"
        ]
      }
    },
...
  }
}